move InvalidTagSequence class to ill_formed.py
